Improve UTF-8 support for c-strings
The default encoding of c-strings shall be UTF-8. C-strings can be
validated by the `cstr_is_utf8` method. There is also a method to
represent c-string with valid UTF-8 content in JSON. However if the
content of the c-string is not UTF-8 compatible the representing method
will use \x escape sequences which are not defined in JSON.